chore: rename countdown components
