+++ src/components/otp/components/OtpCountdown.tsx
@@ -1,13 +1,13 @@
 import { DButton } from '@dynamic-framework/ui-react';
 import { Trans, useTranslation } from 'react-i18next';
 
-import useCountdown from '../hooks/useCountDown';
+import useCountdown from '../hooks/useCountdown';
 
 type Props = {
   seconds: number;
 };
 
-export default function OtpCountDown(
+export default function OtpCountdown(
   {
     seconds,
   }: Props,
